Rename io.py and add dolfinx.io as submodule in setup.py
